+++ ReleaseNotes.md
@@ -14,5 +14,10 @@ November, 2018
 
 - Fix: Volume Snapshot Policy
 
+## What is New in vSphereNetAppToolkit 0.2.0
+November, 2018
+
+- New Function: Get-VsphereNetappVolume
+
 ### Feedback
 Please send your feedback to https://github.com/mycloudrevolution/vSphereNetAppToolkit/issues